cpp: Reduce discovery timeout in tests
During tests, if discovery fails, the timeout is waited before the test
fails. Currently this timeout is quite huge (40 secs). So if a test fails
during arbitration it takes a lot of time until the test ends.

Now the timeout is set to 1 second, because all the test run locally.

Change-Id: I98e22a6c885521b3f64d33e8111868a2d953fc9d