license: Add license and notice files to packaged jar source files
Use the maven-dependency-plugin to extract the LICENSE file to the root
build folder of the maven projects. With the maven build-helper plugin, the
license and notice files are added to the sources and test-sources. This
step is needed by the maven source plugin to generate the source packages.
The execution is done during the prepare-package phase and so during
compile phase this files are not yet part of the sources and therefore not
copied to the build folder. This is done with the maven resource plugin.
It copies the files to:
1. target/classes
2. target/test-classes
3. target/apidocs

This build steps need to be excluded from the android build, because they
produce build errors.

Change-Id: Ia4111b08f8627a2283f2800ce66db25cd76e10be