cpp: Joynr messaging stub uses message sender
Until now, the joynr message sender used the communication manager to send
messages. The communication manager itself delegated the call to a message
sender.

With this patch the joynr messaging stub directly uses the message sender
to send messages.

For mocking purposes a IMessageSender interface is introduced.

The initialization of the message sender is still done by the communication
manager. The message sender is then exported using the getMessageSender
function. This will be changed in the following patches. The redundant send
functionality within the communication manager will also be removed.

Change-Id: I0a664978fae4d136bfc1e7aab8c774f337cbc800