discovery ui: show channel Id
This is a workaround. Currently the capdir is 
transforming the capability information to cap
entries. They should remain as cap info though,
and this change would then not be necessary.This 
will be added to a new ticket for persisting cap 
info on the server (to be changed all at once)

Change-Id: Id916215a550819a98b6afdf7ab65ab95208336ad