Added support for base64 blobs.
The code generators were changed so that fields with the Franca type
'ByteBuffer' have the C++ type QByteArray and the Java type byte[].

The C++ implementation serializes the blobs as base64. The Java
Jackson json parser handles byte[] fields as base64 automatically.

At the moment there is no support for ByteBuffer arguments to methods.

The hash function used an incorrect variable.

Change-Id: I3765c7b053461a0658a4edacecf025fc8c6f27b5