Added HTTP connect timeout to message settings.
The current long poll timeout covers everything from DNS lookup to the download of data. When sending
large messages the long poll timeout could be set to a large value (e.g 5 minutes).  If the long poll
timeout is large then long polling does not recover quickly from connection errors.

The connect timeout allows a developer to set a shorter timeout when opening a connection to
the bounce proxy and be able to receive large messages.

Change-Id: I510e33404da8f09c1d6d261197fbeff79771e0f4