mvn: Restructure activation of java checks
With the change to maven 3.2.2 the build-resouces project in the tools
module causes a cyclic dependency. The cause are the profiles for activating and deactivation specific plugins.

The profiles are specified on the root project so dependencies to the build-resources project introduce the cyclic dependency.

Now the profiles are pushed down one level and deactivated in the tools project, so no cyclic dependency is introduced.

Change-Id: Ie79c2e4f9bbc9f966e69042dafbaaa5fabffc484