[cpp] Register broadcast output parameters as composite metatype
This is needed for correctly interpreting a broadcast publication on
consumer side.

Broadcast output parameters are packed into a single publication message when the
broadcast occurs. Hence, a new composite data type is needed for all broadcasts with
more than one out parameter. This composite type (containing all out parameters) is then
serialised into the publication message. The single output parameter types are registered
in the Qt meta type system. When deserialising on consumer side, the right publication
interpreter is chosen by calculating a type id for the composite type.

Change-Id: Id019203e7d674c2cf03e9a93b92c98980328fb87