java: split capabilities store from cache
In order to maintain the logic in the local capability directory
for dealing with local and remote capabilities, the entries have to be
stored in separate cache and store. When the store is migrated to JPA,
a separate class will be needed to save the remote data in its own
table.

Change-Id: I3b6e89c48f77deaab804e1e9804c4de9b47532ab