[Java] Resend messages on connection error
HttpMessageSender catches an Exception when connection failures
occur. Previously, the messages were dropped. This change causes
messages to be resent.

The change handles a race condition on servlet start where the
Discovery Directory and Global Domain Access controller are
unable to communicate with each other because the HTTP port is
not always open.

Change-Id: I36d44c805fcbea03ee1e0213cd1564a2c754f35c