[Java] use statically provisioned access control in Radio
Start the access control.war when starting the radio-app servers
Preconfigure an access control entry for the current domain
and vehicle/radio interface used in the demo.

NOTE: the provisioned ACEs are currently provisioned only to
the local domain access controller, as the global controller
is already started on the command line using maven. This does
not affect the workings of the demo, but does cause the ACE
to be found locally as opposed to forcing a remote lookup.

Change-Id: I0ac9c4f7e79e9d82e799b2666127458ba0e76dfd