[C++] Reply object provides multiple responses
Prepare reply object to be ready for multiple output parameters of
methods.

Avoid the usage of QList::append method, as in case output parameters
are arrays, the append method does not add the array to the response
object of the reply, but adds all list entries to the response object.
Tested this behaviour with a JsonSerializerTest.

Change-Id: I4765f4472b13e319454eca4a67fe69d38ced08a2