[C++] Adapt constructor API of LocalDiscoveryAggregator
The LocalDiscoveryAggregator expects a discovery proxy implementing
the IDiscoverySync interface. Depending of the aggregator has been
initialised at cc or libjoynr side, the actual provider or a created
discovery proxy has been provided to the aggregator.
This patch now prepares the LocalDiscoveryAggregator, so he can work
also with async providers (introduced in subsequent patches). This is
done, by creating on cc side a DiscoveryInProcessConnector, which is
linked with the real provider and implementing the sync interface.

Change-Id: I1f33d05223aab20013afe35fa61bb70f610a1cef