[Cpp] Remove Qt from DiscoveryQos
removed Qt datatypes from use in DiscoveryQos.

The keyword field was removed from QosArbitrator at the same
time because an exception thrown by std::map.at() in DiscoveryQos
caused creation of the arbitrator to fail. keyword was likely
a copy-and-paste error in the first place, since it is not used
by this arbitrator at all.

Change-Id: I6ef3fc9d3b6462298e7ca5621ccbe5021f750d6e