[model] don't use the interface name as type collection
As interface extends type collection, the interface name was being
used in enums and structs declared within an interface. Once the
type collection name is used in the package name, this causes a
compiler error in Java on interfaces that are all lower-case.

Now not using the interface name in the package name

Change-Id: Ie473a9b85a59da267cde16a924b6a64d4981ee60