[C++] C++ build builds the whole mvn reactor
This is required to ensure the cpp relevant mvn modules use its most
recent dependencies. E.g., if changes in a backend service and c++
code happen within the same patch, it might be the case that the c++
build relies on the most recent change of the backend service.

Originally, this patch has been demanded as a change in the super pom
(dependency management) did not affect the c++ build, causing
unresolved depdendencies errors during the generation step.

Change-Id: Iae7b2197a42667bae54a7415e134af2b18347ec1