[Java] Delay publication transmission in case of bursts
If two attribute updates arrive within the min_interval of the
subscription, the PublicationManager previously dropped the newly
arriving publication. With this patch, the publication is postponed,
but checking at the new time of publication, if the publication has been
become obsolete due to intermediate publications. This situation is
possible in a multi-threaded environment.

Change-Id: Ia3c2d144b1e0b3f7db4620454b5b6f1888e52381