[JS] Get it working to have an attribute as enum array
This needs to change the meta information generated in the
<Interface>Proxy and <Interface>Proxy. The attribute type for lists
shall not be "TypesEnum.LIST" in this case, but the type name of the
enum + "[]”. For migration purposes, the helper function in JSTypeUtil
still allows to generate the old "TypesEnum.LIST" (via the flag
migrateToNewListRepresentation). This will be
removed subsequently.

Added SystemIntegrationTest showing the proper functionality of attributes as enum array.

Change-Id: Icd9e53d4d5c18f5f22fdc7612e4e003eff1d7f5f