[JS] remove test that was causing side effects
The JS builds are no longer running all tests if this test is active.
Unknown why there is a side-effect with ProxyAttribute tests.
Fails only seen in Phantom
Tested if there was a bug / side-effect in the Promise lib bluebird
but was unable to prevent the failures by:
- rewriting Promise.reject to new Promise() and rejecting manually
- resolving instead of rejecting

Also tried rewriting test without using the done() async test method
to no effect.

Change-Id: Idada15d483842aae5b80cd4d76366efbc35c1e81