[Java] Get Radio-App working
Migrate to new way of instantiating the JoynrInjectorFactory

By default, the InProcessRuntime is used having own CC for each
Radio instance.

If an additional parameter "websocket" is given when
invoking the application, the radio app communicates via
websocket with an existing radio cluster controller.

For the MyRadioProviderApplication, it is additionally possible
to use the additionalParameter websocketCC. In this case, the
provider application uses the runtime which starts the cluster
controller listening for web socket connection. This can be used
by the MyRadioConsumerApplication as web socket cluster
controller.

Change-Id: Iea8c3c92e2115bb5f3d6ec0ab87e51d0d6516452