[Java] Radio app allows to configure the provider scope
The provider application accepts as additional
argument "local" saying, if the provider should only be registered
local to its own cluster controller or registered globally.

In addition, the consumer application accepts as well an additional
argument "local". If set, the the arbitration only looks locally
in the cluster controller capabilities directory, if a matching
provider is registered.

Change-Id: Ic8dc41b80636b15fd385221d2dd1c14f3b28c3e4