[C++] Cleanup handling of subscriptionListeners in AbstractJoynrProvider
If no listener for the attribute is registered, the call to
attributeListeners[attributeName] is prevented, ensuring that the
attributeListeners map is NOT touched.

In addition, the unregisterAttributeListener erases the attribute
from the attributeListeners map, in case no listener is attached
to the attribute after this unregister. This ensures a clean map.

The same goes for broadcastListeners.

Change-Id: I9d42c99ec9ab56adc955fba788f0e94230ad1dd1