[JS, Tools] Do not use TypesEnum in generated code
Do not use util class "TypesEnum" to specify the metadata of method
parameters in the generated. Instead, hard code the type information
for primitive types in the generated proxies and providers. This way,
the generic template function "joynrTypeName" can be directly re-used.

In a subsequent step, the util class TypesEnum can then be completely
removed.

Change-Id: I68b398c0d8581b10c95f0a08676755bf689789ed