[C++] Fix correct invocation of broadcast filters for array params
Broadcast filters have been invoked incorrectly in case the
output parameters of broadcasts were arrays. This patch enriches
an End2EndBroadcastTest with a filter and checks its correct invocation.

The required adaption in the filter template is provided by this patch
as well.

This patch also checks for enum arrays as broadcast parameter

Change-Id: I003d0db1a6082dc7a55b8b283fb9c0f46f0f2c12