[JS] Distinguish DiscoveryEntry and CapabilityInformation in Runtime
Info: the return parameter of the GlobalCapabilitiesDirectory.lookup is
a capability information, while the return parameter of the
Discovery.lookup is a DiscoveryEntry.

The provisioned/default capabilities are always of type
"CapabilityInformation", and are transformed into a
"DiscoveryEntry" if appropriate.

Note for InProcessRuntime and InterTabClusterControllerRuntime:
All the user provisioned capabilities are seen as global capabilities.
Thus, they are now added to the global capabilities cache.

Change-Id: I001da61132a3f686f8687708a803e8f3ff9db05b