[C++,Test] Introduce simple application for joynr system int tests
The consumer and provider here implemented are communicating through
the SystemIntegrationTest interface via websocket.

The consumer performs a test sequence and exits with EXIT_SUCCESS in
case the provider replies as expected. As of now, the conumser
invokes the "add" method and checks if the result is the sum of the two
in arguments.

The provider exits once all the expected test sequences have been
performed. As of now, he expects the call to the "add" method.

Change-Id: Ief68f00dc95e4f6c8adc0b13530e7f1fbfca11ce