[C++] avoid slow generation of UUIDs
Analysis showed that the UUID generation causes a constant overhead
of ~0.7 ms upon creation of each JoynrMessage.

The instantiation of the random generator which is used for UUID
generation is expensive, therefore a static random generator is now used
which avoid this overhead as the actual generation of the UUID is
optimized very well.

Change-Id: Iefe5f0c23f994b9e17d1eba03a7cb0442dcd97ec