[C++, Test] Use different participant IDs in SystemServicesRoutingTest.
Anticipate fixes for the persistency of MessageRouter.

At test startup the messageRouter will be reloaded from persistency.
Tests included in SystemServicesRoutingTest are looking for a specific participant IDs (which will
be already there if it was used in another test case).

This commit makes sure a new participant id is generated for each new test case.

Change-Id: Ic3da782daa60b7664f34177861499ce3ca76ca8f