[C++] Moved connector selection to proxy builder
This patch is a preparation for a subsequent patch which will remove
the CommunicationMiddleware type. The corresponding enumeration was
used to select either an in-process or messaging connector during the
creation process of a proxy. Now, the selection is done by using a
boolean value which is determined directly within the proxy builder and
forwarded to the connector factory. This eliminates the need for the
CommunicationMiddleware type.

Change-Id: I95899f4b7b09451ac14c63b1c4b0789499285522