[Java] Make the channelUrl available to the GlobalAddressFactory
In preparation for the ChannelAddress to also contain the endpoint
Url: the HttpGlobalAddressFactory needs to know the url, but this
url is first known after the channel is created on the bounceproxy.
Solution: allow the factory to register as a listener to be notified
of the url once the channel is created.

Change-Id: Id7a2ec4ecde4525ddfb102283c82e85dd589218a