[Java, API] Refactor GlobalAddressProvider
GlobalAddressProvider needs to provide the "best" WAN address from
the multibinder set.

* Instead of address knowledge existing in the provider, the "best"
address is either the only one (if one only) or is marked with the
property "joynr.messaging.primaryglobaltransport"
* Since GlobalAddressFactories may not know the global address at
injection time (the case now for ChannelAddress, since the endpoint
URL is not set until the channel is created on the bounceproxy), a
new listener concept was introduced to inform the LocalCapabilities
Directory of the global address to be used in the GlobalDiscoveryEntry
address field once it becomes available.
* If the global address is not available, the LocalCapabilitiesDirectory
queues discoveryEntries with scope GLOBAL.

Change-Id: Ie712eddbbeead0558289647ccc3c59f9495f9dbc