[JeeIntegration] Enrich <Interface>Sync with annotations
The annotation provides info about the implementing provider of this
interface (@ProvidedBy). In addition, information is shared about the
proxy (@UsedBy).

These annotations can be enabled by setting the parameter "jee"
in the configuration of the joynr-generator-maven-plugin. Have a look
at java/jeeintegration/pom.xml for an example of how to use it.

The joynr standalone generator is also able to process the command
line argument "-jee true".

Change-Id: Ic0df338ed34bbe808c14b6a0fbf18ba2ce6ed2ab