[C++] Moved disconnect callback out of WebSocketMessagignStubFactory
WebSocketMessagingStubFactory created a 'disconnect' lambda callback
function which used the 'this' pointer and was finally passed to a
websocket. The problem is, that there is no guarantee that the callback
will not be used after the WebSocketMessagignStubFactory object was
deleted.

Therefore, the callback creation was moved to LibJoynrWebSocketRuntime
and WebSocketCCMessagingSkeleton because both can access the
WebSocketMessagignStubFactory as a shared_ptr.

Change-Id: I4f5b0cc1b50f97fe1fd30f6630bbe94658f97c80