[JEE] Allow JEE apps to use MQTT for incoming messages
A new config option allows enabling the HTTP bridge option, which
requires an MQTT broker with a plugin which can forward incoming MQTT
messages via HTTP to endpoints registered with a registry available as a
REST service (see JEE integration documentation).
The default is for the bridge NOT to be enabled.

In order for the receiving of messages via MQTT to work, we have to use a
version of Paho which allows us to pass in an executor service which is
then used for threading. Currently this is a fork of paho at
https://github.com/bmwcarit/paho.mqtt.java
Once the changes are upstreamed to the main paho
project, we will switch back to that.

Change-Id: Idf61184c58573e8f22074af36bf75e3baf2b08a5