[ILT] split attributeWithException into setter and getter exception
AttributeWithException has to be split into attributeWithExceptionFromGetter
and attributeWithExceptionFromSetter to be able to test failing getter calls
and failing setter calls.

The attribute setter implementation in ProviderAttribute.js calls
the getter and setter implementations to automatically trigger
attribute publications if the attribute value has changed.

Since the getter is called first and the execution is stopped on error,
the exception from the getter implementation is returned instead of the
wanted exception from the setter implementation.

Change-Id: I7808e1d174f066c35a4eaa24a65dcee66cef0853