[JEE, Examples] example radio jee app
An example implementation of the radio app using the JEE integration.

The implementation can also be used as a guide for how to structure
a joynr app being implemented for JEE.

Note that because jetty will start up on port 8080 by default, and
Glassfish / Payara also do, copied and changed the joynr backend services
config from radio-app to start jetty on 8383 instead.
The jee apps are also configured to connect to that port.

In joynr 0.18 there is a known bug whereby JEE apps will not start up if
the endpoint registry url is not configured. This is fixed in 0.19 and
there doesn't need to be specified at all in the configs if the http
bridge is not being used.

The JEE integration does not currently support subscription / broadcasts,
yet the radio.fidl includes both. In order to be able to successfully
deploy the application, the provider needs to at least have the
SubscriptionPublisherInjection interface, because the joynr runtime expects
to be able to perform the injection. Added as an empty method with a comment
pointing out that the JEE integration does not currently support this.

Change-Id: I1b69ecd8f927714900ed37423cbb22803e49741c