[Java] StaticCapabilitiesProvisioning now fails early
Previously the static cap prov was only logging errors occuring during JSON parsing
and simply leaving the discovered entries as an empty set. This makes it a bit
harder to realise that something's gone wrong. Hence, this was changed to fail
early - the class will throw a JoynrRuntimeException and stop Guice in its tracks
if the user has configured invalid JSON for the capabilities.

Change-Id: Icdf866f9030abbf5afeba1fad6dc521e2c536982