[C++] Sync add & lookup calls of LocalCapabilitiesDirectory
In case a client performs a lookup with scope LOCAL_THEN_GLOBAL,
it might be that the call is waiting for the result of the global
capabilities directory call. While waiting, a new capability might be
registered locally to the LocalCapabilitiesDirectory. In this case,
the LocalCapabilityDirectory should check if such an incomplete lookup
call is matching with the domain + interface name of the added discovery
entry. In case of a match, the incomplete lookup should be completed
by returning the newly added capability.

Change-Id: Ibd6b7aaf6c72830ab8dc4536856c3f5c2f17df98