[Java] made mqtt client ID unique
For clustered scenarios the topic in the address is not unique
because all cluster nodes share it. Hence, we need to generate
as truly a unique ID as we can by using the receiver ID
(which is set via generating a random UUID and is then persisted,
so we know it will be the same next time the app starts).

Change-Id: Idac35d5950da0e613e7bf7980a2affca116b92f2