[Java] ChildMessageRouter: Don't use Runnable when not threading
Runnables were being used to store addNextHop requests when the parent
router was unknown. This was confusing, as these stored requests were
then processed synchronously in setParentRouter, not by a thread
executor or similar as the use of Runnable would suggest.

It was also completely unnecessary, as the only data stored in the
runnable was a String (participantId).

Change-Id: Iba29dd0ba81d1deab61093dcb141e705ee5b23f9