[Java] multi-out containers marked with interface
In order to easily identify a container class generated for multi-out
return values we add a marker interface which the generated class
implements.

The interface and the generated container class provide a method which
allows a caller to obtain the array of values representing the
multi-out result. This can then be used to resolve a deferred
instance with the correct values.

Change-Id: I44fcdd3bb6c806b192ad9299f2a8081321f2d480