[Java, JEE] use the multi-out interface to create the correct deferred
When resolving deferred instances for multi-out methods, we need to
resolve using the array of values held in the multi-out container
instance rather than the container itself. In order to do this
we introduce a MultiValueDeferred class which passes the varargs
straight through to the abstract deferred, and in the provider
wrapper we check if the result we are working with is an instance
of the MultiReturnValuesContainer, and give that extra treatment
if so.

Change-Id: I39af66c572e2d2acc1c8c4c1a552e46d9f3fba5d