[Java] refactor message creation to only pass MessagingQos
When calling the message factory, the callers were basically splatting out
the content of the MessagingQos into individual parameters rather than
simply passing in the MessagingQos. Now they only pass the one parameter.
This also makes it more future proof for when we add more content to
the messaging QoS class.

Change-Id: Ibbec15e737e5c58a6801b0bc28893f56d21d0171