[JEE] add beans implementing JoynrMessageProcessor as processors
Any EJB which implements the JoynrMessageProcessor interface (and only that interface)
and can be found by the default joynr runtime factory via the bean manager will be
added as a joynr message processor to the JoynrMessageProcessorProvider. Then, any time
a message is created via the JoynrMessageFactory, the processors found and added are
called.

The calling is actually proxied, because we can't pass in direct references to bean instances.
Instead, a proxy is passed in, which knows the bean and the bean manager. It then obtains
a bean instance and calls the process method on that.
Of course, this means, that it can only be called from a thread which is within the EE
execution context (i.e. one spawned by a managed executor service, for example.

Change-Id: I2998f91bd6d37c960b266957d4e70c80616f4c80