[JEE] bug fix for JEE Radio App configuration
With the change that capabilities are provisioned via a JSON file and the
legacy properties are still supported, but not the primary means of configuration
we had to add the participant IDs to the configs. Also changed the persistence
file name for joynr.properties in each so that they don't overwrite each others
config when deployed to the same JEE container.

Change-Id: I244afafd4ad0288c9b1e523bcecd2d2b26bad5aa