[Java] defaulted participant IDs for cap dir and acl dir
So that you can customise the address of the statically provisioned
capability and domain access controller directories by only providing
the URLs, the participant IDs have default values (again). Now it is
only an error if you configure a custom URL via properties, but set the
participant ID to null / empty.

Change-Id: I49e712d0abd9fb7a2e27f5755558d88d87b6898e