[Java] static capabilities can be loaded from a URI resource
Previously the value specified for PROPERTY_PROVISIONED_CAPABILITIES_FILE had
to be a file path or a path to a classpath resource. Now it can also be a
URI, such as an http:// URL, in order to be able to provide the JSON file
for provisioning from a web resource (either statically served content or
even a dynamic endpoint which calculates the values to use for the given
infrastructure environment).

Change-Id: I98446ba688d658b881871378e2cc3267e1c70edd