[Java, JEE] changed provider of JoynrMessageProcessors to multibinder
Instead of the singleton JoynrMessageProcessorProvider which provides
a list of JoynrMessageProcessors, we changed it to use a Guice multibinder
which provides a set of JoynrMessageProcessors. If we need ordering in
the future, then we'll add another message to the interface
JoynrMessageProccessor which allows the implementer to provide a
weighting for the interface, and then sort the set of processors
before using them.

Change-Id: Ia93fc3e70521ddc64482518ed33db5bc51e4e42f