[Java] change expired cleanup to use a callback
Instead of the expired discovery entry cache cleaner removing the entries which
have expired itself, we pass in a callback which is called for all the expired
entries. The user of the cleaner component can then decide itself exactly how
the entries are to be cleaned up.
We also changed that to be by calling the local capabilities directory's
remove(DiscoveryEntry) method instead of simply removing the entry from the
relevant DiscoveryEntryStore.

Change-Id: Ie7241434626ec775252b8b4cfe6772b2c7fd3bd9