[JEE] set subscription publisher on all bean instances
Previously the subscription publisher was only being set on the bean instance
being created for the call to the setSubscriptionPublisher method, which was
wrong because more than one bean can potentially be created by the ProviderWrapper.
Hence, we store the subscription publisher created for the registered provider
in the ProviderWrapper and set it on each bean instance obtained from the
BeanManager for all subsequent calls, hence ensuring it's available to the
beans if joynr created one.

More detail: an instance of the provider wrapper is used as a proxy
for the provider which is registered with the joynr runtime. If joynr
now wants to call the provider, the wrapper is invoked, which will then
ask the JEE CDI BeanManager for an EJB instance to actually perform
the provider logic.
Potentially this could be a new EJB instance every time. Hence, if we
only set the subscription publisher on the first bean we obtain this
way, all subsequent EJB instances would not have the subscription
publisher available.

Change-Id: I005f73573f26a41930b1990ae9856b648c27982b