[JEE] enable subscription publishers to be injected
In order for the concrete subscription publisher instances to be
injected in CDI beans, we need to change the injection type to
the base type SubscriptionPublisher so that we can provide a
producer method for that type, which then uses the meta data
of the injection points to determine the actual type and also
that we have a directory mapping bean classes to the actual
instances to be used. In a later step the integration can
then store the real subscription publishers in that directory
at the point they are provided by the joynr runtime.

Change-Id: I717188bc208b6e170fe9cd0782002349369666ea