[C++] Fix issue with infinite sendMessages loop
If a messaging stub is removed (this happens if a
disconnect occurs) while the MessageRouter is sending queued
message for the stub's address, the while(true) loop will
never terminate. The reason is that sendMessage tries to
create a messaging stub and reinserts the message into the
queue because the creation fails.

Change-Id: I6e84c1b255898ee8bf0512e12623f36f49c1ec99