[JEE] runtime factory creates clusterable participant IDs
Because JEE applications can be deployed in a cluster, the
provider participant IDs need to be the same for each runtime
of each node within a given cluster. In order to ensure this
the JEE integration's DefaultJoynrRuntimeFactory will create
clusterable participant IDs for each provider which is
provisioned in the application being deployed.

Change-Id: I1afbce15e01aa07fb64efe5d4d311c57d4ea1092