[JS] Correct deserialisation of error enum objects
Previously, the JS deserialisation code assumed that error enums
coming via a joynr message are complete in the sense
that it contains _typename, value and name information.

However, customer experience has shown that this is not the case in
all joynr bindings (JS, C++, Java). Thus, this patch reduces this
constraint and assumes only _typename and name as being present in
the joynr message. Code walktrough has shown that this is information
is provided in all three languages.

Change-Id: I6f8b4d70f62a6c0ef64f9756622c1efa5b3d65cc