[JS] Simplify own appenders
The appenders to not need to add info to the formatted message
like log level, logger name a.s.o. This is already covered by the
layout manger. E.g. the PatternLayout allows to add timestamp info
(%d), logger name (%c) or log level (%p) to the log output. In addition,
it is possible to specify to number of levels to be expanded for
objects as log messages (%m{<level>}).

More details about the PatternLayout can be found here:
http://log4javascript.org/docs/manual.html#patternlayout

Change-Id: I6c72b06cab9eef2d012cdaf03d17c7edd90dbac9