[JS] Fix incorrect handling with timeout ids
The very first missed reply timer of the RequestReplyManager has never
been cleared in case of successful reply, as it had the id 0
and the if condition "if (0)" return false. Thus, this patch fixes
twice the mentioned error. First, the if condition is now
"if (0 !== undefined)", second, the id of timers starts with 1.

The same behaviour has been identified in ChannelMessagingSender.

Checked also in other classes (SubscriptionManager, PublicationManager,
SharedWebSocket, TypeRegistry) for improper check condition.

Change-Id: I0da1f312b5dacabbfd02dd262d726252770701a9