[JS] Fix memory leak in LongTimer.setTimeout
For intervals less than 2^31-1 the native setTimeout was called
directly and an entry added to timeoutMap, which was then never
removed. Using pure proxy calls of setTimeout and clearTimeout is
beneficial for both CPU and memory performance.

Changing the internal behavior of LongTimer.setTimeout emphasised
a bug in ChannelMessagingSender, as it showed a potential circular
dependency. This is fixed now, as the setTimeout function of node
returns an object which contains circular references, and
ChannelMessagingSender tried to serialise this object.

Change-Id: Id0998521b0bda381bea76a9cf5731bf31b296665