[JS] Exclude _typeName from map checkMembers and move to prototype
Do not include _typeName when checking type of values in the map,
as _typeName is always a String, while the keys can be complex types.

Also moved the function out of the instance and into the prototype
so that it works the same as the checkMembers from the complex types.

Change-Id: I830e656c7a8fafaf0e75babba7970027b346ab87