[JS] move node_modules created in browserify out of main directory
node_modules was being duplicated in the libjoynr-js directory and in
libjoynr-js/target/node-classes directory. Having a node_modules in
libjoynr-js caused problems when trying to run karma from the cli
as not all the joynr dependencies were installed in the duplicate.
Instead merged the duplicate with node_modules in the node-classes
directory and then can create a symbolic link from the main directory
to target/node-classes/node_modules to run karma

Change-Id: I6115414f430f92cf983f6cff75a566bb583be10e