[Java] Make RequestCaller independent to prepare for setting context
RequestCaller was previously an empty interface used solely for
proxying calls to the provider object. This prevented it from being
expanded to allow setting the context on its provider, as all method
calls on a RequestCaller were proxied to the invoke method.

Now RequestCaller contains a separate proxy object that it used to
forward method calls to, while being able to offer its own methods as
well.

The tests were modified to use Test.fidl as the fake TestRequestCaller
objects were no longer working now that RequestCaller is not a mere
proxy.

Change-Id: I29d69d6234f8267a2c85a8f629afb0e2a96e9acb