[Java] Move runtime cleanup in ProviderProxyEnd2EndTests
The runtime was cleaned up by storing the corresponding joynr application
and calling cleanup on this object. However, it is easy to miss this
part. Therefore store the created runtimes and clean them up in the
base class of the tests. Then the getRuntime method does not have to
take care of registering the created application / runtime.

Change-Id: I97ca64f196b4985580c511ee0a6bc936bb62301f