[Maven] Do not call joynr-generator-maven-plugin while cleaning
This is not needed as the maven-clean-plugin is already configured to
delete the folders created by the generator in the "generate" phase. So
after calling "mvn clean" the generated files are deleted even without
calling the joynr-generator-maven-plugin. Furthermore, this avoids the
need for maven to get the joynr-generator-maven-plugin when simply
calling "mvn clean".

Change-Id: I1164f855a949c6fcc0fe8adae6bd86ea3c7e72a6