[Java, Maven] Add slf4j-api dependency to generator framework
This change is needed in order to have slf4j-api on the classpath of the
generator framework. Reflections 0.9.8 has declared an optional
dependency to slf4j-api, however the library throws an NPE when there is
no logger and this can break the joynr-generator-maven-plugin. The issue
is solved in newer versions of the reflections library but we are not
able to update by now.

Change-Id: Icce0bfd5e245ce48e9f84b987821c715c33150c6