[C++] fix mapping of websocket connection to ownerId of certificate
The ssl verify callback (getCNFromCertificate) which extracts the
ownerId of the TLS certificate is called for every certificate of
the certificate chain starting with the ca certificate.
The ownerId of the last certifcate of the certificate chain
(= client certificate) has to be stored in the clients map.
Map::emplace does not work because the former former mapping has
to be replaced.

Change-Id: I5275aeeb8f3b4747308eca1f7b01140755d841fb