[Java] Do not shutdown runtimes twice
The provider and the consumer runtimes are added to the list of
createdRuntimes. This way they could be shut down with all
createdRuntimes and would not need an extra shutdown call. However,
the order of shutdown is important in joynr in order to get it
exception-free (e.g. shut down CC last). That is why the separate calls
are left in the code but both runtimes are removed from the
createdRuntimes list.

Change-Id: I44edb503cf0a1179dba2c7cc86cd7768b78f5a58