[Java] Stop MessageWorkers the Java way
Before this patch the threads (from the pool) running the
MessageWorkers were forcefully stopped using interrupts. It was not
clear at what state the MessageWorker is stopped. Now gracefully stop
the MessageWorkers "informing" them and allow them to stop in a
controlled way. This allow the threads (from the pool) running the
MessageWorks to stop gracefully as well.

Change-Id: Ie623a27961a47ad5bd084ece62f8eedd2cc13881