[JS] perf: ProviderEvent use prototype
Earlier it was the case that each new object would not only create
new variables but also new functions on each object creation. This
is unnecessary, because those functions can be reused. By writing
those functions on the prototype of the object, they will automatically
be inherited by a new object creation.

In order to keep the same encapsulation a proxy Object was created
which has the same interface but won't show its internal implementation
and therefore external manipulation is avoided.

Change-Id: I4681c2f2564bc49f5e39404ccebabb28e161a135