[JS] add forwardPrototype to UtilInternal
ForwardPrototype is a helper function. It is like Util.forward
which was intruduced earlier.
ProxyEvent, ProxyOperation and ProxyAttribute, ProviderEvent,
ProviderOperation and ProviderAttribute all shared a common
problem. They did not use prototypes but created all their
functions once again every time their constructor was called.
This can of course bloat the memory. But introducing prototype
is not so simple in this case.
There are 2 issues. First of all Object.freeze was called on
the this object which meant that it could not be manipulated
any more. Secondly it is important to make sure that the
functions are called with the correct context.
Here exactly comes forwardPrototype into play. This functionality
abstracts the process of forwarding the public functions to
a new dummy Object and binding them there, to make sure the
right context is being used. That way Object.freeze() can
still be used and the correct context is no Problem because
the functions are bound.

Change-Id: Iec647eca2413aa632dcd2acc467d66a35d66d385