[JS] joynrMessage use smrfMsg internally but accept both externally
There are 2 use cases when a new joynrMessage is being created.
First: joynr creates a message in the dispatcher which  will be send
later. Second: joynr receives a message, which was deserialized from
smrf. Since js only allows for one constructor, this is actually
a dilemma. When a joynrMessage is being created in the dispatcher,
arguments are continously added (e.g effort, custom headers etc.)
Therefore the function joynrMessage.parseMessage was added. Since
the earlier constuction with overriding the prototype of the settings
object didn't acutally use the joynrMessage instance, it was logical
to move this kind of constuction externally to the function
JoynrMessage.parseMessage. Now manipulation for the __proto__
object won't be necessary when the dispatcher creates the message
and this is a performance improvement because this is a bailout
for v8 optimizations.

The second part of the commit is the change of the internal stucture
of joynrMessage. This was less difficult than expected because when
accessing e.g. joynrMessage.expiryDate joynrMessage.header.expiryDate
was actually accessed. This works with the help of
Object.defineProperty. Since there was already a mapping, this mapping
could be changed to use the internal structure of smrfMessage.
When interaction with joynrMessage.expiryDate now, joynrMessage.ttlMs
is being read or written. This makes it possible to replace the
mapping process which was executed in MessageSerializer before and
the joynrMessage may simply be passed to smrf to serialize.

Change-Id: I397425117545740782fc80f0360e458302f39283